+++ examples/example-search-api/pages/_app.tsx
@@ -1,5 +1,20 @@
+import * as React from "react"
+import {
+  Hydrate,
+  QueryClient,
+  QueryClientProvider,
+} from "@tanstack/react-query"
+
 import "tailwindcss/tailwind.css"
 
 export default function App({ Component, pageProps }) {
-  return <Component {...pageProps} />
+  const [queryClient] = React.useState(() => new QueryClient())
+
+  return (
+    <QueryClientProvider client={queryClient}>
+      <Hydrate state={pageProps.dehydratedState}>
+        <Component {...pageProps} />
+      </Hydrate>
+    </QueryClientProvider>
+  )
 }